frontend overhaul with tailwind implementation
